// packages/aridhia-client/src/generated/apis/AuthenticationApi.ts
/* tslint:disable */
/* eslint-disable */
/**
 * C-Path FAIR API (Filtered)
 * Filtered C-Path FAIR API containing only required endpoints
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import * as runtime from '../runtime'
import type {
  AuthenticationRequest,
  AuthenticationResponse,
} from '../models/index'
import {
  AuthenticationRequestFromJSON,
  AuthenticationRequestToJSON,
  AuthenticationResponseFromJSON,
  AuthenticationResponseToJSON,
} from '../models/index'

export interface AuthenticatePostRequest {
  authenticationRequest: AuthenticationRequest
}

/**
 *
 */
export class AuthenticationApi extends runtime.BaseAPI {
  /**
   * Exchange third party token
   */
  async authenticatePostRaw(
    requestParameters: AuthenticatePostRequest,
    initOverrides?: RequestInit | runtime.InitOverrideFunction,
  ): Promise<runtime.ApiResponse<AuthenticationResponse>> {
    if (requestParameters['authenticationRequest'] == null) {
      throw new runtime.RequiredError(
        'authenticationRequest',
        'Required parameter "authenticationRequest" was null or undefined when calling authenticatePost().',
      )
    }

    const queryParameters: any = {}

    const headerParameters: runtime.HTTPHeaders = {}

    headerParameters['Content-Type'] = 'application/json'

    if (this.configuration && this.configuration.accessToken) {
      const token = this.configuration.accessToken
      const tokenString = await token('bearerAuth', [])

      if (tokenString) {
        headerParameters['Authorization'] = `Bearer ${tokenString}`
      }
    }
    const response = await this.request(
      {
        path: `/authenticate`,
        method: 'POST',
        headers: headerParameters,
        query: queryParameters,
        body: AuthenticationRequestToJSON(
          requestParameters['authenticationRequest'],
        ),
      },
      initOverrides,
    )

    return new runtime.JSONApiResponse(response, jsonValue =>
      AuthenticationResponseFromJSON(jsonValue),
    )
  }

  /**
   * Exchange third party token
   */
  async authenticatePost(
    requestParameters: AuthenticatePostRequest,
    initOverrides?: RequestInit | runtime.InitOverrideFunction,
  ): Promise<AuthenticationResponse> {
    const response = await this.authenticatePostRaw(
      requestParameters,
      initOverrides,
    )
    return await response.value()
  }
}