// packages/aridhia-client/src/generated/apis/RequestsApi.ts
export interface FairRequestsCodeGetRequest {
  code: string
}

export interface FairRequestsGetRequest {
  page?: number
  pageSize?: number
}

export interface FairRequestsPostRequest {
  requestPost: RequestPost
}

/**
 *
 */
export class RequestsApi extends runtime.BaseAPI {
  /**
   * Get a specific request by code
   */
  async fairRequestsCodeGetRaw(
    requestParameters: FairRequestsCodeGetRequest,
    initOverrides?: RequestInit | runtime.InitOverrideFunction,
  ): Promise<runtime.ApiResponse<Request>> {
    if (requestParameters['code'] == null) {
      throw new runtime.RequiredError(
        'code',
        'Required parameter "code" was null or undefined when calling fairRequestsCodeGet().',
      )
    }

    const queryParameters: any = {}

    const headerParameters: runtime.HTTPHeaders = {}

    if (this.configuration && this.configuration.accessToken) {
      const token = this.configuration.accessToken
      const tokenString = await token('bearerAuth', [])

      if (tokenString) {
        headerParameters['Authorization'] = `Bearer ${tokenString}`
      }
    }
    const response = await this.request(
      {
        path: `/fair/requests/{code}`.replace(
          `{${'code'}}`,
          encodeURIComponent(String(requestParameters['code'])),
        ),
        method: 'GET',
        headers: headerParameters,
        query: queryParameters,
      },
      initOverrides,
    )

    return new runtime.JSONApiResponse(response, jsonValue =>
      RequestFromJSON(jsonValue),
    )
  }

  /**
   * Get a specific request by code
   */
  async fairRequestsCodeGet(
    requestParameters: FairRequestsCodeGetRequest,
    initOverrides?: RequestInit | runtime.InitOverrideFunction,
  ): Promise<Request> {
    const response = await this.fairRequestsCodeGetRaw(
      requestParameters,
      initOverrides,
    )
    return await response.value()
  }

  /**
   * List all requests
   */
  async fairRequestsGetRaw(
    requestParameters: FairRequestsGetRequest,
    initOverrides?: RequestInit | runtime.InitOverrideFunction,
  ): Promise<runtime.ApiResponse<FairRequestsGet200Response>> {
    const queryParameters: any = {}

    if (requestParameters['page'] != null) {
      queryParameters['page'] = requestParameters['page']
    }

    if (requestParameters['pageSize'] != null) {
      queryParameters['pageSize'] = requestParameters['pageSize']
    }

    const headerParameters: runtime.HTTPHeaders = {}

    if (this.configuration && this.configuration.accessToken) {
      const token = this.configuration.accessToken
      const tokenString = await token('bearerAuth', [])

      if (tokenString) {
        headerParameters['Authorization'] = `Bearer ${tokenString}`
      }
    }
    const response = await this.request(
      {
        path: `/fair/requests/`,
        method: 'GET',
        headers: headerParameters,
        query: queryParameters,
      },
      initOverrides,
    )

    return new runtime.JSONApiResponse(response, jsonValue =>
      FairRequestsGet200ResponseFromJSON(jsonValue),
    )
  }

  /**
   * List all requests
   */
  async fairRequestsGet(
    requestParameters: FairRequestsGetRequest = {},
    initOverrides?: RequestInit | runtime.InitOverrideFunction,
  ): Promise<FairRequestsGet200Response> {
    const response = await this.fairRequestsGetRaw(
      requestParameters,
      initOverrides,
    )
    return await response.value()
  }

  /**
   * Create a new request
   */
  async fairRequestsPostRaw(
    requestParameters: FairRequestsPostRequest,
    initOverrides?: RequestInit | runtime.InitOverrideFunction,
  ): Promise<runtime.ApiResponse<Request>> {
    if (requestParameters['requestPost'] == null) {
      throw new runtime.RequiredError(
        'requestPost',
        'Required parameter "requestPost" was null or undefined when calling fairRequestsPost().',
      )
    }

    const queryParameters: any = {}

    const headerParameters: runtime.HTTPHeaders = {}

    headerParameters['Content-Type'] = 'application/json'

    if (this.configuration && this.configuration.accessToken) {
      const token = this.configuration.accessToken
      const tokenString = await token('bearerAuth', [])

      if (tokenString) {
        headerParameters['Authorization'] = `Bearer ${tokenString}`
      }
    }
    const response = await this.request(
      {
        path: `/fair/requests/`,
        method: 'POST',
        headers: headerParameters,
        query: queryParameters,
        body: RequestPostToJSON(requestParameters['requestPost']),
      },
      initOverrides,
    )

    return new runtime.JSONApiResponse(response, jsonValue =>
      RequestFromJSON(jsonValue),
    )
  }

  /**
   * Create a new request
   */
  async fairRequestsPost(
    requestParameters: FairRequestsPostRequest,
    initOverrides?: RequestInit | runtime.InitOverrideFunction,
  ): Promise<Request> {
    const response = await this.fairRequestsPostRaw(
      requestParameters,
      initOverrides,
    )
    return await response.value()
  }
}